React navigation > folder refactoring
