+++ src/screens/subjects/subject_details.tsx
